move some files + add about
