+++ src/components/AboutPage/About.js
@@ -1,0 +1,32 @@
+import React from 'react';
+
+const About = () => {
+	return (
+		<div className=' flex justify-center min-w-[300px]'>
+			<pre>
+				{`
+Welcome to TeamsGenerator, your ultimate solution for effortless team formation!
+
+Say goodbye to the headache of manually organizing players into teams. With TeamsGenerator,
+you can create balanced and engaging teams with just a few clicks.
+
+Why Choose Us?
+
+Efficiency: Save time and effort by letting our algorithm do the work for you.
+
+Fairness: Enjoy competitive yet fair teams based on skill levels and preferences.
+
+Flexibility: Customize team size and preferences to suit your needs.
+
+User-Friendly: Our intuitive interface makes team formation a breeze for everyone.
+
+Engaging Experience: Focus on the fun while we handle the teams.
+
+Join us in redefining team building. Experience seamless team formation with TeamsGenerator today! 
+(made by chatGPT)`}
+			</pre>
+		</div>
+	);
+};
+
+export default About;
