+++ src/components/TeamGeneratorPage/TeamCompCard.js
