+++ src/components/TeamGeneratorPage/common/TeamCompCard.js
