+++ src/components/TeamGeneratorPage/services/generateTeams.js
@@ -14,14 +14,11 @@ export function pickTeams(list) {
 
 export function getStoredPlayersList() {
 	const list = localStorage.getItem('player_list');
-	console.log('get', list);
 	if (JSON.stringify(list).length > 2) {
 		return list.split(',');
 	}
 	return [];
 }
 export function setStoredPlayersList(players) {
-	console.log('set', players);
-
 	localStorage.setItem('player_list', players);
 }
